+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/full.php
@@ -24,6 +24,7 @@ $container->loadFromExtension('framework', array(
     'session' => array(
         'storage_id'      => 'session.storage.native',
         'handler_id'      => 'session.handler.native_file',
+        'on_demand_mode'  => 1,
         'name'            => '_SYMFONY',
         'cookie_lifetime' => 86400,
         'cookie_path'     => '/',